Initial

+++ Enemy.h
@@ -1,18 +1,17 @@
-  #include "Object.h"
-  #include "EventCollision.h"
-  #include "EventView.h"
+#include "Object.h"
+#include "EventCollision.h"
+#include "EventView.h"
 
 #define FISH_ENEMY "enemy-fish"
 #define SEAWEED_ENEMY "enemy-seaweed"
 
 class Enemy : public df::Object {
 
-   public:
+public:
     Enemy(std::string spriteName);
     ~Enemy();
-    int eventHandler(const df::Event *p_e);
-    void moveToStart(int y);
+    int eventHandler(const df::Event* p_e);
+    void moveToStart(float y);
     void out();
-    void hit(const df::EventCollision *p_c);
-
-  };
+    void hit(const df::EventCollision* p_c);
+};
